+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.6] - 2026-04-28
+
+### Docs
+- Update README.md
+
+### Other
+- Update regres/doctor.py
+- Update regres/import_error_toon_report.py
+- Update regres/regres_cli.py
+
 ## [0.1.5] - 2026-04-28
 
 ### Docs